Move some files to bfabric.results module
